Rename api to api_c
