+++ libvcx/src/api_c/logger.rs
