+++ libvcx/src/api_c/return_types_u32.rs
