+++ libvcx/src/api_c/utils.rs
@@ -619,7 +619,7 @@ mod tests {
 
     use agency_client::mocking::AgencyMockDecrypted;
 
-    use crate::api::return_types_u32;
+    use crate::api_c::return_types_u32;
     use crate::utils::constants;
     use crate::utils::devsetup::*;
     use crate::utils::timeout::TimeoutUtils;
