+++ libvcx/src/disclosed_proof.rs
@@ -233,7 +233,7 @@ mod tests {
     extern crate serde_json;
 
     use serde_json::Value;
-    use crate::api::VcxStateType;
+    use crate::api_c::VcxStateType;
     use crate::aries::messages::proof_presentation::presentation_request::PresentationRequestData;
     use crate::utils::constants::{ARIES_PROVER_CREDENTIALS, ARIES_PROVER_SELF_ATTESTED_ATTRS, GET_MESSAGES_DECRYPTED_RESPONSE};
     use crate::utils;
