+++ libvcx/tests/logging.rs
@@ -7,11 +7,11 @@ extern crate futures;
 
 use self::libc::{c_void, c_char};
 use std::ptr::null;
-use vcx::api::logger::*;
+use vcx::api_c::logger::*;
 use vcx::utils::logger::{LOGGER_STATE, LoggerState};
 use indy::wallet;
 use vcx::utils::cstring::CStringUtils;
-use vcx::api::logger::vcx_set_logger;
+use vcx::api_c::logger::vcx_set_logger;
 
 /// These tests can only be run individually as initing the log crate can happen
 /// only once.
@@ -20,7 +20,7 @@ use vcx::api::logger::vcx_set_logger;
 /// logging is outputting to stdout.
 mod log_tests {
     use super::*;
-    use vcx::api::vcx::vcx_error_c_message;
+    use vcx::api_c::vcx::vcx_error_c_message;
     use indy::future::Future;
 
     static mut COUNT: u32 = 0;
